+++ web/src/GameScreen.js
@@ -2,26 +2,9 @@ import React, { useState, useEffect, useContext } from 'react';
 import { Context } from './context.js';
 import {Grid, Paper, Typography, Button, TextField, makeStyles} from '@material-ui/core';
 
-const useStyles = makeStyles(theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  paper: {
-    padding: theme.spacing(2),
-    textAlign: 'center',
-    color: theme.palette.text.secondary,
-    height: 500
-  },
-  img: {
-    maxHeight: '100%',
-    maxWidth: '100%'
-  }
-}));
-
-function GameScreen({setScreen}) {
+function GameScreen(setScreen, classes) {
   const [context, setContext] = useContext(Context)
-  const classes = useStyles();
-
+  console.log(classes); 
   return (
     <div className={classes.root}>
       <Grid container
